feature/ASE-01: renamed folders for the change logs, force move party event step now has a movement speed, use floats in global message script step

+++ Kurumi-Engine/Game/Scripts/Context/Capabilities/Interfaces/Universal/IUIActions.cs
@@ -14,7 +14,7 @@ public interface IUIActions : ICapability
 
     public IFinishable OpenTextWindowWithNameBox(IReadOnlyList<string> pages, string name);
 
-    public void OpenGlobalMessage(int timeLimit, string text);
+    public void OpenGlobalMessage(float timeLimit, string text);
 
     public ChoiceBoxWithDialogueOverlay OpenTextWindowWithChoice(IReadOnlyList<string> choices, string text);
 }
